[#4267] Refactor API endpoints and related tests
