+++ src/openforms/contrib/zgw/clients/catalogi.py
@@ -31,15 +31,13 @@ class CatalogiClient(NLXClient):
         all_data = pagination_helper(self, data)
         return list(all_data)
 
-    def get_all_informatieobjecttypen(
-        self, *, catalogus: str | None = None
-    ) -> list[dict]:
+    def get_all_informatieobjecttypen(self, *, catalogus: str = "") -> list[dict]:
         """List all informatieobjecttypen.
 
         :arg catalogus: the catalogus URL the informatieobjecttypen should belong to.
         """
         params = {}
-        if catalogus is not None:
+        if catalogus:
             params["catalogus"] = catalogus
         response = self.get("informatieobjecttypen", params=params)
         response.raise_for_status()
