+++ src/openforms/registrations/api/urls.py
@@ -1,11 +1,6 @@
 from django.urls import include, path
 
-from .views import (
-    AllAttributesListView,
-    CatalogiListView,
-    InformatieObjectTypenListView,
-    PluginListView,
-)
+from .views import AllAttributesListView, PluginListView
 
 urlpatterns = [
     path("plugins", PluginListView.as_view(), name="registrations-plugin-list"),
@@ -14,18 +9,7 @@ urlpatterns = [
         AllAttributesListView.as_view(),
         name="registrations-attribute-list",
     ),
-    path(
-        "catalogi",
-        CatalogiListView.as_view(),
-        name="catalogi-list",
-    ),
-    path(
-        "informatieobjecttypen",
-        InformatieObjectTypenListView.as_view(),
-        name="iotypen-list",
-    ),
 ]
-
 
 # add plugin URL patterns
 # TODO: make this dynamic and include it through the registry?
@@ -34,5 +18,9 @@ urlpatterns += [
     path(
         "plugins/objects-api/",
         include("openforms.registrations.contrib.objects_api.api.urls"),
+    ),
+    path(
+        "plugins/zgw-api/",
+        include("openforms.registrations.contrib.zgw_apis.api.urls"),
     ),
 ]
