+++ src/openforms/registrations/contrib/zgw_apis/api/urls.py
@@ -1,0 +1,18 @@
+from django.urls import path
+
+from .views import CatalogiListView, InformatieObjectTypenListView
+
+app_name = "zgw_apis"
+
+urlpatterns = [
+    path(
+        "catalogi",
+        CatalogiListView.as_view(),
+        name="catalogi-list",
+    ),
+    path(
+        "informatieobjecttypen",
+        InformatieObjectTypenListView.as_view(),
+        name="iotypen-list",
+    ),
+]
